requirements.txt
beautifulsoup4==4.13.3
decorator==5.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
markdownify==1.1.0
moviepy==2.1.2
numpy==2.2.3
pillow==10.4.0
proglog==0.1.10
PyPDF2==3.0.1
python-dotenv==1.0.1
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
